Add build script for osslsigncode
Based on https://github.com/sporsh/docker-osslsigncode/blob/master/Dockerfile
It downloads osslsigncode source from sourceforge and builds the binary.

The build script adds the local workdir as a volume and overrides the binary
to commit.

Note: This version is has libgsf linked as dependency to also support msi signing on linux.